Moved custom log formatter out of initializers, so that it isn't initialized twice.
