+++ config/application.rb
@@ -1,7 +1,7 @@
 # frozen_string_literal: true
 
-# require_relative 'boot'
-require_relative 'initializers/log_format'
+require_relative 'boot'
+require_relative 'log_format'
 require 'rails/all'
 
 # Require the gems listed in Gemfile, including any gems
